# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jelajah Lokal API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/articles": {
      "get": {
        "summary": "List Articles",
        "operationId": "list_articles_api_admin_articles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_list_ArticleDraft__"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Article",
        "operationId": "create_article_api_admin_articles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_ArticleDraft_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/articles/{article_id}": {
      "get": {
        "summary": "Get Article",
        "operationId": "get_article_api_admin_articles__article_id__get",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Article Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_ArticleDraft_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Article",
        "operationId": "update_article_api_admin_articles__article_id__patch",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Article Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticlePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_ArticleDraft_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Article",
        "operationId": "delete_article_api_admin_articles__article_id__delete",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Article Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_Deleted_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/articles/{article_id}/publish": {
      "post": {
        "summary": "Publish Article",
        "operationId": "publish_article_api_admin_articles__article_id__publish_post",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Article Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_ArticleDraft_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/articles/{article_id}/unpublish": {
      "post": {
        "summary": "Unpublish Article",
        "operationId": "unpublish_article_api_admin_articles__article_id__unpublish_post",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Article Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_ArticleDraft_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/categories": {
      "get": {
        "summary": "List Categories",
        "operationId": "list_categories_api_admin_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_list_Category__"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Category",
        "operationId": "create_category_api_admin_categories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_Category_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/categories/{category_id}": {
      "get": {
        "summary": "Get Category",
        "operationId": "get_category_api_admin_categories__category_id__get",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_Category_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Category",
        "operationId": "update_category_api_admin_categories__category_id__patch",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Category Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_Category_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Category",
        "operationId": "delete_category_api_admin_categories__category_id__delete",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_Deleted_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/authors": {
      "get": {
        "summary": "List Authors",
        "operationId": "list_authors_api_admin_authors_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_list_Author__"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Author",
        "operationId": "create_author_api_admin_authors_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_Author_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/authors/{author_id}": {
      "get": {
        "summary": "Get Author",
        "operationId": "get_author_api_admin_authors__author_id__get",
        "parameters": [
          {
            "name": "author_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Author Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_Author_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Author",
        "operationId": "update_author_api_admin_authors__author_id__patch",
        "parameters": [
          {
            "name": "author_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Author Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_Author_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Author",
        "operationId": "delete_author_api_admin_authors__author_id__delete",
        "parameters": [
          {
            "name": "author_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Author Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_Deleted_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/media": {
      "get": {
        "summary": "List Media",
        "operationId": "list_media_api_admin_media_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_list_Media__"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Upload Media",
        "operationId": "upload_media_api_admin_media_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_media_api_admin_media_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_Media_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/media/{media_id}": {
      "delete": {
        "summary": "Delete Media",
        "operationId": "delete_media_api_admin_media__media_id__delete",
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_Deleted_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArticleCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "title": "Title",
            "default": ""
          },
          "slug": {
            "type": "string",
            "maxLength": 120,
            "title": "Slug",
            "default": ""
          },
          "body": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Paragraph"
                },
                {
                  "$ref": "#/components/schemas/Heading"
                },
                {
                  "$ref": "#/components/schemas/Quote"
                },
                {
                  "$ref": "#/components/schemas/ListBlock"
                },
                {
                  "$ref": "#/components/schemas/ImageBlock"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "heading": "#/components/schemas/Heading",
                  "image": "#/components/schemas/ImageBlock",
                  "list": "#/components/schemas/ListBlock",
                  "paragraph": "#/components/schemas/Paragraph",
                  "quote": "#/components/schemas/Quote"
                }
              }
            },
            "type": "array",
            "maxItems": 300,
            "title": "Body"
          },
          "excerpt": {
            "type": "string",
            "maxLength": 240,
            "title": "Excerpt",
            "default": ""
          },
          "metaDescription": {
            "type": "string",
            "maxLength": 160,
            "title": "Metadescription",
            "default": ""
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Tags"
          },
          "categoryDocumentId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categorydocumentid"
          },
          "authorDocumentId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authordocumentid"
          },
          "coverMediaDocumentId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Covermediadocumentid"
          },
          "coverAlt": {
            "type": "string",
            "maxLength": 180,
            "title": "Coveralt",
            "default": ""
          },
          "coverCaption": {
            "type": "string",
            "maxLength": 240,
            "title": "Covercaption",
            "default": ""
          },
          "imageCredit": {
            "type": "string",
            "maxLength": 240,
            "title": "Imagecredit",
            "default": ""
          },
          "imageSourceUrl": {
            "type": "string",
            "maxLength": 2048,
            "title": "Imagesourceurl",
            "default": ""
          },
          "sourceLinks": {
            "items": {
              "$ref": "#/components/schemas/SourceLink"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Sourcelinks"
          },
          "informationCheckedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Informationcheckedat"
          },
          "operationalClaims": {
            "type": "boolean",
            "title": "Operationalclaims",
            "default": false
          },
          "regionLabel": {
            "type": "string",
            "maxLength": 80,
            "title": "Regionlabel",
            "default": ""
          },
          "featured": {
            "type": "boolean",
            "title": "Featured",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ArticleCreate"
      },
...
      "AuthorInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 2,
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "maxLength": 120,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Slug"
          },
          "bio": {
            "type": "string",
            "maxLength": 300,
            "title": "Bio",
            "default": ""
          },
          "avatarMediaDocumentId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatarmediadocumentid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "slug"
        ],
        "title": "AuthorInput"
      },
...
      "CategoryInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2,
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "maxLength": 120,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Slug"
          },
          "description": {
            "type": "string",
            "maxLength": 240,
            "title": "Description",
            "default": ""
          },
          "order": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": -10000.0,
            "title": "Order",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "slug"
        ],
        "title": "CategoryInput"
      },
...
      "Data_ArticleDraft_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ArticleDraft"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Data[ArticleDraft]"
      },
...
      "Data_Category_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Category"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Data[Category]"
      },
      "Data_Deleted_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Deleted"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Data[Deleted]"
      },
...
      "Data_list_ArticleDraft__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ArticleDraft"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Data[list[ArticleDraft]]"
      },
      "Data_list_Author__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Author"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Data[list[Author]]"
      },
      "Data_list_Category__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Category"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Data[list[Category]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RevisionInput": {
        "properties": {
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision"
        ],
        "title": "RevisionInput"
      },
...
    }
  }
}